[Codechange] RigSpawner: control flow code moved to separate file (easier debugging)
